feat:moved to ts + Abstract intersection observer to another component
